refactoring: form shell submodule (#5550)
#5538
