+++ pyproject.toml
@@ -27,7 +27,8 @@ readme = {file = ["README.rst"]}
 packages = [
     "xonsh",
     "xonsh.ply.ply",
-    "xonsh.ptk_shell",
+    "xonsh.shell",
+    "xonsh.shell.ptk_shell",
     "xonsh.procs",
     "xonsh.platform",
     "xonsh.parsers",
@@ -242,11 +243,11 @@ convention = "numpy"
 "xonsh/style_tools.py" = ["F821"]
 "xonsh/xoreutils/*.py" = ["E722"]
 "xonsh/completers/python.py" = ["E722"]
-"xonsh/ptk/shell.py" = ["E731"]
+"xonsh/shell/ptk_shell/__init__.py" = ["E731"]
 "xonsh/ast.py" = ["F401"]
-"xonsh/readline_shell.py" = ["F401"]
+"xonsh/shell/readline_shell.py" = ["F401"]
 "xonsh/commands_cache.py" = ["F841"]
-"xonsh/ptk/key_bindings.py" = ["F841"]
+"xonsh/shell/ptk_shell/key_bindings.py" = ["F841"]
 "xonsh/tools.py" = [
     "E731",
 ]
