+++ tests/shell/test_base_shell.py
@@ -1,11 +1,11 @@
-"""(A down payment on) Testing for ``xonsh.base_shell.BaseShell`` and associated classes"""
+"""(A down payment on) Testing for ``xonsh.shell.base_shell.BaseShell`` and associated classes"""
 
 import os
 
 import pytest
 
-from xonsh.base_shell import BaseShell
 from xonsh.shell import transform_command
+from xonsh.shell.base_shell import BaseShell
 
 
 def test_pwd_tracks_cwd(xession, xonsh_execer, tmpdir_factory, monkeypatch):
