+++ tests/shell/test_ptk_highlight.py
