+++ tests/shell/test_shell.py
