+++ xonsh/environ.py
@@ -1535,7 +1535,7 @@ class PromptSetting(Xettings):
         "`prompt_toolkit <https://github.com/jonathanslenders/python-prompt-toolkit>`_"
         " library installed. To specify which shell should be used, do so in "
         "the run control file. "
-        "It also accepts a class type that inherits from ``xonsh.base_shell.BaseShell``",
+        "It also accepts a class type that inherits from ``xonsh.shell.base_shell.BaseShell``",
         doc_default="``best``",
     )
     SUGGEST_COMMANDS = Var.with_default(
