+++ xonsh/pytest/plugin.py
@@ -363,7 +363,7 @@ def ptk_shell(xonsh_execer):
     from prompt_toolkit.input import create_pipe_input
     from prompt_toolkit.output import DummyOutput
 
-    from xonsh.ptk_shell.shell import PromptToolkitShell
+    from xonsh.shell.ptk_shell import PromptToolkitShell
 
     out = DummyOutput()
     with create_pipe_input() as inp:
@@ -375,7 +375,7 @@ def ptk_shell(xonsh_execer):
 
 @pytest.fixture
 def readline_shell(xonsh_execer, tmpdir, mocker):
-    from xonsh.readline_shell import ReadlineShell
+    from xonsh.shell.readline_shell import ReadlineShell
 
     inp_path = tmpdir / "in"
     inp = inp_path.open("w+")
