+++ xonsh/pytest/tools.py
@@ -12,7 +12,7 @@ from collections import defaultdict
 
 import pytest
 
-from xonsh.base_shell import BaseShell
+from xonsh.shell.base_shell import BaseShell
 
 VER_MAJOR_MINOR = sys.version_info[:2]
 VER_FULL = sys.version_info[:3]
