+++ xonsh/shell/ptk_shell/formatter.py
@@ -4,7 +4,7 @@ import functools
 import typing as tp
 
 from xonsh.prompt.base import DEFAULT_PROMPT, PromptFormatter
-from xonsh.ptk_shell.updator import AsyncPrompt, PromptUpdator
+from xonsh.shell.ptk_shell.updator import AsyncPrompt, PromptUpdator
 
 
 class PTKPromptFormatter(PromptFormatter):
