+++ xonsh/shell/ptk_shell/key_bindings.py
